+++ deployments/bin/pre_upgrade.sh
@@ -1,0 +1,4 @@
+# This file is always executed before serenity upgrade.
+# Put below the pre-upgrade instructions like in the upgrade.md
+
+echo "No pre-upgrade instructions."